Module popupRegistration.js was created
