+++ js/script.js
@@ -1,4 +1,6 @@
 import handleBurgerButtons from "./modules/handleBurgerButtons.js";
 handleBurgerButtons();
 
+import popupRegistration from "./modules/popupRegistration.js";
+popupRegistration();
